backend/api/requirements.txt
asgiref==3.10.0
cffi==2.0.0
cryptography==46.0.3
Django==5.2.8
djangorestframework==3.16.1
pillow==12.0.0
pycparser==2.23
PyMySQL==1.1.2
sqlparse==0.5.3
tzdata==2025.2
